+++ SteamManager/Services/IGameLibraryService.cs
@@ -1,12 +1,8 @@
 namespace SteamManager.Services;
 
-/// <summary>
-/// Interface for retrieving the user's Steam game library.
-/// </summary>
 public interface IGameLibraryService
 {
-    /// <summary>
-    /// Gets all games owned by the current Steam user.
-    /// </summary>
     Task<List<Models.GameInfo>> GetOwnedGamesAsync();
+    Task<List<Models.GameInfo>> GetCachedGamesAsync();
+    Task SaveGamesCacheAsync(List<Models.GameInfo> games);
 }